refactor: EncourageFragment로 이름 변경
