+++ app/src/main/java/com/sdhong/jonbeowin/feature/encourage/EncourageFragment.kt
@@ -1,4 +1,4 @@
-package com.sdhong.jonbeowin.feature.encouragingword
+package com.sdhong.jonbeowin.feature.encourage
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.sdhong.jonbeowin.databinding.FragmentEncouragingWordBinding
 
-class EncouragingWordFragment : Fragment() {
+class EncourageFragment : Fragment() {
 
     private var _binding: FragmentEncouragingWordBinding? = null
     private val binding get() = _binding!!
