+++ app/src/main/res/layout/fragment_encourage.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.encouragingword.EncouragingWordFragment">
+    tools:context=".feature.encourage.EncourageFragment">
 
     <TextView
         android:layout_width="match_parent"
